fix (weather): end of month bug fixed
The five day forecast was previously getting the current day of month incrementing it for each day. This meant that at the end of the month, the current day of the month could be 35 which is invalid.

The loop now uses the setDate() function instead to wrap to the next month automatically.