fix(weather): Current high and low now more accurate
Previously the high and low for the current day were mistakenly taken from api data that represented a margin of error for current temperature, not actual highs and lows.

The datapoints now come from actual temperature datapoints given by the api from the current time to the end of the day. This is not super accurate but is the best thing available for this free tier api.